[696] fix: split control-plane registry into modules
